fix: remove visual node
also removes an unneeded readme
